Project1/requirements.txt
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
torch==2.3.1
torchvision==0.18.1a0
pillow==10.0.0
seaborn==0.12.2
pandas==2.0.3
joblib==1.3.1
tqdm==4.66.1 